renamed component name to landingPageFooter
